clips provider refactor
